Separate PokemonDatabase script
